Add useWalletConnect and useRegisterSafeContext hooks for improved modularity
- Introduced `useWalletConnect` and `useRegisterSafeContext` hooks in a new `walletConnect.ts` file to encapsulate WalletConnect context logic.
- Refactored the `WalletConnectProvider` to export the context, enhancing modularity and adherence to Vite's Consistent Component Exports rule.
- Updated the `walletconnect.tsx` route to utilize the new hooks, streamlining the integration with WalletConnect.

These changes improve code organization and maintainability, facilitating easier access to WalletConnect functionalities across the application.
